requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-multipart==0.0.6
docxtpl==0.16.7
python-docx==1.1.0
